Refactor clickhouse itcase.
