+++ bitsail-test/bitsail-test-integration/bitsail-test-integration-base/src/test/java/com/bytedance/bitsail/test/integration/FakeToPrintITCase.java
@@ -26,6 +26,6 @@ public class FakeToPrintITCase extends AbstractIntegrationTest {
   @Test
   public void testFakeToPrint() throws Exception {
     BitSailConfiguration jobConf = JobConfUtils.fromClasspath("fake_to_print.json");
-    submit(jobConf);
+    submitJob(jobConf);
   }
 }
