+++ bitsail-test/bitsail-test-integration/bitsail-test-integration-base/src/test/resources/fake_to_print.json
@@ -4,7 +4,7 @@
       "cid": 0,
       "domain": "test",
       "job_id": -23,
-      "job_name": "bitsail_integration_test_fake_to_print",
+      "job_name": "bitsail_test_integration_fake_to_print",
       "instance_id": -203,
       "user_name": "root",
       "bitsail_location": "./"
