+++ bitsail-test/bitsail-test-integration/bitsail-test-integration-clickhouse/src/test/java/com/bytedance/bitsail/test/integration/clickhouse/container/ClickhouseContainerHolder.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.bytedance.bitsail.connector.clickhouse;
+package com.bytedance.bitsail.test.integration.clickhouse.container;
 
 import com.clickhouse.client.ClickHouseClient;
 import com.clickhouse.client.ClickHouseCredentials;
@@ -54,20 +54,20 @@ public class ClickhouseContainerHolder {
 
   private static final String DROP_TABLE_SQL = "DROP TABLE IF EXISTS " + TABLE;
   private static final String CREATE_TABLE_SQL =
-          "CREATE TABLE " + TABLE + "\n" +
-                  "(\n" +
-                  "    `id`            `Int64`,\n" +
-                  "    `int_type`      `Int32`,\n" +
-                  "    `double_type`   `Float64`,\n" +
-                  "    `string_type`   `String`,\n" +
-                  "    `p_date`        `Date`,\n" +
-                  "    `int_128`       `Int128`,\n" +
-                  "    `int_256`       `Int256`\n" +
-                  ") ENGINE = MergeTree()\n" +
-                  "PARTITION BY toYYYYMM(p_date)\n" +
-                  "PRIMARY KEY (id);";
+      "CREATE TABLE " + TABLE + "\n" +
+          "(\n" +
+          "    `id`            `Int64`,\n" +
+          "    `int_type`      `Int32`,\n" +
+          "    `double_type`   `Float64`,\n" +
+          "    `string_type`   `String`,\n" +
+          "    `p_date`        `Date`,\n" +
+          "    `int_128`       `Int128`,\n" +
+          "    `int_256`       `Int256`\n" +
+          ") ENGINE = MergeTree()\n" +
+          "PARTITION BY toYYYYMM(p_date)\n" +
+          "PRIMARY KEY (id);";
   private static final String INSERT_SQL_HEADER = "INSERT INTO " + TABLE
-          + " (id, int_type, double_type, string_type, p_date, int_128, int_256) VALUES ";
+      + " (id, int_type, double_type, string_type, p_date, int_128, int_256) VALUES ";
   private static final String COUNT_SQL = "SELECT count(id) FROM " + TABLE;
   private static final String SCAN_SQL = "SELECT id, int_type, double_type, string_type, p_date, int_128, int_256 from " + TABLE + " ORDER BY id";
   private static final int INSERT_BATCH_SIZE = 10;
