+++ bitsail-test/bitsail-test-integration/bitsail-test-integration-clickhouse/src/test/resources/example_data.csv
